most of the work done on the new SurfaceList
