+++ scopesim/optics/radiometry_utils.py
@@ -143,16 +143,17 @@ def add_surface_to_dict(dic, surf, name, position=0):
 
 
 def make_surface_dict_from_table(tbl):
-    names = tbl[real_colname("name", tbl.colnames)]
     surf_dict = OrderedDict({})
-    for ii in range(len(tbl)):
-        surf_dict[names[ii]] = make_surface_from_row(tbl[ii], **tbl.meta)
+    if tbl is not None and len(tbl) > 0:
+        names = tbl[real_colname("name", tbl.colnames)]
+        for ii in range(len(tbl)):
+            surf_dict[names[ii]] = make_surface_from_row(tbl[ii], **tbl.meta)
 
     return surf_dict
 
 
 def make_surface_from_row(row, **kwargs):
-    row_dict = {colname.lower() : row[colname] for colname in row.colnames}
+    row_dict = {colname.lower(): row[colname] for colname in row.colnames}
     kwargs.update(row_dict)
     surface = SpectralSurface(**kwargs)
